Revert "Refactor controllers"
