+++ Gemfile
@@ -11,8 +11,6 @@ gem 'sinatra-activerecord'
 gem 'sqlite3'
 gem 'sinatra-flash'
 gem 'jwt'
-gem 'omniauth-cognito-idp'
-gem 'omniauth-oauth2'
 
 group :development, :test do
   gem 'pry'
