+++ db/schema.rb
@@ -18,4 +18,5 @@ ActiveRecord::Schema[7.1].define(version: 1) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
+
 end
