+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
       context: ./
       dockerfile: Dockerfile
     ports:
-      - "3000:4567"
+      - "4567:4567"
     volumes:
       - .:/app
       - rubygems:/usr/local/bundle
